+++ backend/canisters/local_group_index/api/src/updates/c2c_create_community.rs
@@ -16,7 +16,6 @@ pub struct Args {
     pub permissions: Option<CommunityPermissions>,
     pub gate: Option<AccessGate>,
     pub default_channels: Vec<String>,
-    #[serde(default)]
     pub default_channel_rules: Option<Rules>,
     pub source_group: Option<SourceGroup>,
     pub primary_language: String,